+++ packages/axios/src/index.ts
@@ -1,21 +1,69 @@
-import { computed } from "@vue/composition-api";
-import axios, { AxiosRequestConfig } from "axios";
-import { usePromise } from "@vue-composable/core";
+import { computed, Ref, ref } from "@vue/composition-api";
+import axios, { AxiosRequestConfig, AxiosResponse, AxiosInstance, CancelTokenSource } from "axios";
+import { usePromise, PromiseResultFactory } from "@vue-composable/core";
 
 /* istanbul ignore next  */
 const _axios = axios || (globalThis && (globalThis as any).axios);
 
-export function useAxios<TData = any>(config?: AxiosRequestConfig) {
+interface AxiosReturn<TData> extends PromiseResultFactory<Promise<AxiosResponse<TData>>, [AxiosRequestConfig]> {
+  readonly client: Ref<Readonly<AxiosInstance>>;
+  readonly data: Ref<TData | null>;
+  readonly status: Ref<number | null>;
+  readonly statusText: Ref<string | null>;
+
+  cancel: (message?: string) => void;
+  readonly isCancelled: Ref<boolean>;
+  readonly cancelledMessage: Ref<string | null>;
+  // readonly 
+}
+
+// TODO replace by shared project
+const isObject = (d: any): d is Object => typeof d === 'object';
+const isString = (d: any): d is String => typeof d === 'string';
+const isBoolean = (d: any): d is Boolean => typeof d === 'boolean';
+
+
+export function useAxios<TData = any>(throwException?: boolean): AxiosReturn<TData>;
+export function useAxios<TData = any>(url: string, config?: AxiosRequestConfig, throwException?: boolean): AxiosReturn<TData>;
+export function useAxios<TData = any>(url: string, throwException?: boolean): AxiosReturn<TData>;
+export function useAxios<TData = any>(config?: AxiosRequestConfig, throwException?: boolean): AxiosReturn<TData>;
+export function useAxios<TData = any>(configUrlThrowException?: AxiosRequestConfig | string | boolean, configThrowException?: AxiosRequestConfig | boolean, throwException = false): AxiosReturn<TData> {
   /* istanbul ignore next  */
   __DEV__ &&
     !_axios &&
     console.warn(`[axios] not installed, please install it`);
 
+  const config = !isString(configUrlThrowException) && !isBoolean(configUrlThrowException) ? configUrlThrowException : isObject(configThrowException) ? configThrowException as AxiosRequestConfig : undefined;
+  throwException = isBoolean(configUrlThrowException) ? configUrlThrowException : isBoolean(configThrowException) ? configThrowException : throwException;
+
   const axiosClient = _axios.create(config);
   const client = computed(() => axiosClient);
+  const isCancelled = ref(false);
+  const cancelledMessage = ref<string>(null);
+
+  let cancelToken: CancelTokenSource | undefined = undefined;
+  const cancel = (message?: string) => {
+    if (!cancelToken) {
+      if (__DEV__) throw new Error('Cannot cancel because no request has been made');
+      return;
+    }
+    cancelToken.cancel(message);
+    isCancelled.value = true;
+    if (message) {
+      cancelledMessage.value = message;
+    }
+  }
+
   const use = usePromise(async (request: AxiosRequestConfig) => {
-    return axiosClient.request(request);
-  });
+    cancelToken = _axios.CancelToken.source()
+    isCancelled.value = false;
+    cancelledMessage.value = null;
+
+    return axiosClient.request<any, AxiosResponse<TData>>({
+      cancelToken: cancelToken.token,
+      ...request
+    });
+  }, throwException);
 
   const data = computed<TData>(
     () =>
@@ -25,7 +73,7 @@ export function useAxios<TData = any>(config?: AxiosRequestConfig) {
         use.error.value.response.data) ||
       null
   );
-  const status = computed(
+  const status = computed<number>(
     () =>
       (use.result.value && use.result.value.status) ||
       (use.error.value &&
@@ -33,7 +81,7 @@ export function useAxios<TData = any>(config?: AxiosRequestConfig) {
         use.error.value.response.status) ||
       null
   );
-  const statusText = computed(
+  const statusText = computed<string>(
     () =>
       (use.result.value && use.result.value.statusText) ||
       (use.error.value &&
@@ -42,11 +90,23 @@ export function useAxios<TData = any>(config?: AxiosRequestConfig) {
       null
   );
 
+  // if url provided in the config, execute it straight away
+  // NOTE: `false` is passed to the `exec` to prevent the exception to be thrown
+  if (typeof configUrlThrowException === 'string') {
+    (use.exec as any)({ ...config, url: configUrlThrowException }, false);
+  } else if (config && config.url) {
+    (use.exec as any)(config, false)
+  }
+
   return {
     ...use,
     client,
     data,
     status,
-    statusText
+    statusText,
+
+    cancel,
+    isCancelled,
+    cancelledMessage
   };
 }