+++ packages/core/__tests__/promise/promise.spec.ts
@@ -102,6 +102,7 @@ describe("promise", () => {
     });
   });
 
+
   it("should update the result when calling multiple times", async () => {
     let result = 0;
     const use = usePromise(factory);
@@ -181,6 +182,106 @@ describe("promise", () => {
       error: { value: null }
     });
   });
+
+  describe('throw exception', () => {
+    it('should throw when throwException is true at creation', async () => {
+      expect.assertions(2);
+
+      const error = new Error("error");
+      const use = usePromise(
+        () => Promise.reject(error),
+        true
+      );
+
+      try {
+        await use.exec();
+      }
+      catch (e) {
+        expect(e).toMatchObject(error);
+      }
+
+      await nextTick();
+
+      expect(use).toMatchObject({
+        result: { value: null },
+        loading: { value: false },
+        error: { value: error }
+      });
+    });
+
+    it('should throw if the last argument of the exec is true, but factory is ()=>void', async () => {
+      expect.assertions(2);
+
+      const error = new Error("error");
+      const use = usePromise(
+        () => Promise.reject(error)
+      );
+
+      try {
+        await (use.exec as any)(true);
+      }
+      catch (e) {
+        expect(e).toMatchObject(error);
+      }
+
+      await nextTick();
+
+      expect(use).toMatchObject({
+        result: { value: null },
+        loading: { value: false },
+        error: { value: error }
+      });
+    });
+
+    it('should throw if the last argument of the exec is true, but factory is (a)=>void', async () => {
+      expect.assertions(2);
+
+      const error = new Error("error");
+      const use = usePromise(
+        (a: number) => Promise.reject(error)
+      );
+
+      try {
+        await (use.exec as any)(1, true);
+      }
+      catch (e) {
+        expect(e).toMatchObject(error);
+      }
+
+      await nextTick();
+
+      expect(use).toMatchObject({
+        result: { value: null },
+        loading: { value: false },
+        error: { value: error }
+      });
+    });
+
+    it('should not throw if the last argument of the exec is false', async () => {
+      expect.assertions(1);
+
+      const error = new Error("error");
+      const use = usePromise(
+        () => Promise.reject(error),
+        true
+      );
+
+      try {
+        await (use.exec as any)(false);
+      }
+      catch (e) {
+        expect(e).toMatchObject(error);
+      }
+
+      await nextTick();
+
+      expect(use).toMatchObject({
+        result: { value: null },
+        loading: { value: false },
+        error: { value: error }
+      });
+    });
+  })
 
   it("should throw if factory is invalid", () => {
     expect(() => {