+++ packages/core/src/debounce.ts
@@ -1,5 +1,5 @@
-export function useDebounce<T extends Function>(handler: T, wait?: number): T {
-  return debounce(handler as any, wait);
+export function useDebounce<T extends Function>(handler: T, wait?: number, options?: Options): T {
+  return debounce(handler as any, wait, options);
 }
 
 /**
@@ -12,7 +12,6 @@ export type Options = {
   isImmediate: boolean;
 };
 
-/* istanbul ignore next */
 export function debounce<F extends Procedure>(
   func: F,
   waitMilliseconds = 50,
@@ -22,10 +21,10 @@ export function debounce<F extends Procedure>(
 ): F {
   let timeoutId: ReturnType<typeof setTimeout> | undefined;
 
-  return function(this: any, ...args: any[]) {
+  return function (this: any, ...args: any[]) {
     const context = this;
 
-    const doLater = function() {
+    const doLater = function () {
       timeoutId = undefined;
       if (!options.isImmediate) {
         func.apply(context, args);