+++ packages/core/src/promise/promise.ts
@@ -1,4 +1,5 @@
 import { ref, Ref } from "@vue/composition-api";
+import { isBoolean } from "../utils";
 
 type PromiseType<T extends Promise<any>> = T extends Promise<infer R>
   ? R
@@ -8,7 +9,7 @@ interface PromiseResult<
   T extends Promise<any>,
   TR = PromiseType<T>,
   TError = any
-> {
+  > {
   promise: Ref<T | undefined>;
   result: Ref<TR | null>;
 
@@ -18,21 +19,51 @@ interface PromiseResult<
 
 export interface PromiseResultFactory<
   T extends Promise<any>,
-  TArgs extends Array<any> = never
-> extends PromiseResult<T> {
+  TArgs extends Array<any> = Array<any>
+  > extends PromiseResult<T> {
   exec: (...args: TArgs) => Promise<PromiseType<T> | undefined>;
 }
 
-// TODO fix the typings, T doesn't need to extend Promise<any>
+/**
+ * 
+ * @param fn - factory function
+ * @param throwException - if `true` allows to catch exception when `exec()`
+ */
+export function usePromise<T = any, TArgs extends Array<any> = Array<any>>(
+  fn: (...args: TArgs) => Promise<T>, throwException?: boolean
+): PromiseResultFactory<Promise<T>, TArgs>;
+
+export function usePromise<T = any, TArgs extends Array<any> = Array<any>>(
+  fn: (...args: TArgs) => Promise<T>): PromiseResultFactory<Promise<T>, TArgs>;
+
+
+export function usePromise<T = any, TArgs extends Array<any> = Array<any>>(
+  fn: (...args: TArgs) => T, throwException: boolean
+): PromiseResultFactory<Promise<T>, TArgs>;
+
+export function usePromise<T = any, TArgs extends Array<any> = Array<any>>(
+  fn: (...args: TArgs) => T): PromiseResultFactory<Promise<T>, TArgs>;
+
+
+export function usePromise<T = any>(
+  fn: () => Promise<T>, throwException: boolean
+): PromiseResultFactory<Promise<T>>;
+
+export function usePromise<T = any>(
+  fn: () => Promise<T>
+): PromiseResultFactory<Promise<T>>;
+
+export function usePromise<T = any>(
+  fn: () => T, throwException: boolean
+): PromiseResultFactory<Promise<T>>;
+
+export function usePromise<T = any>(
+  fn: () => T
+): PromiseResultFactory<Promise<T>>;
+
 
 export function usePromise<T extends Promise<any>, TArgs extends Array<any>>(
-  fn: (...args: TArgs) => T
-): PromiseResultFactory<T, TArgs>;
-export function usePromise<T extends Promise<any>>(
-  fn: () => T
-): PromiseResultFactory<T>;
-export function usePromise<T extends Promise<any>, TArgs extends Array<any>>(
-  fn: (...args: TArgs) => T
+  fn: (...args: TArgs) => T, throwException = false
 ): PromiseResultFactory<T, TArgs> {
   if (!fn) {
     throw new Error(`[usePromise] argument can't be '${fn}'`);
@@ -51,8 +82,11 @@ export function usePromise<T extends Promise<any>, TArgs extends Array<any>>(
     error.value = null;
     result.value = null;
 
+    let throwExp = args && fn.length !== args.length && args.length > 0 && isBoolean(args[args.length - 1]) ? args[args.length - 1] : throwException;
+
     const currentPromise = (promise.value = fn(...args));
     try {
+
       const r = await currentPromise;
       if (promise.value === currentPromise) {
         result.value = r;
@@ -63,7 +97,7 @@ export function usePromise<T extends Promise<any>, TArgs extends Array<any>>(
         error.value = er;
         result.value = null;
       }
-      return undefined;
+      return throwExp ? currentPromise : undefined;
     } finally {
       if (promise.value === currentPromise) {
         loading.value = false;